project/requirements.txt
flask==2.3.3
easyocr==1.7.0
pandas==2.0.3
openpyxl==3.1.2
pillow==10.0.0
numpy==1.24.3
werkzeug==2.3.7
python-dotenv==1.0.0
PyPDF2==3.0.1
# For frontend animation (install via npm):
# npm install framer-motion